Integrate Web Control Panel directly into primary launcher NexusTradingForexBot.py
- Updated NexusTradingForexBot.py to serve as the unified launcher, concurrently starting both the web control dashboard (on an available TCP port starting at 8080) and the quantitative AI trading engine on the same asyncio event loop.
- Kept main.py in sync by delegating execution directly to NexusTradingForexBot.py to prevent code duplication.
- Ensured absolute coherence between the modern dark-theme user interface and the backend database, state, positions, and AI predictions.
- Verified that all unit and integration tests pass perfectly.
